+++ modules/backup/outputs.tf
@@ -21,3 +21,12 @@ output "vm_backup_policy_id" {
   value       = azurerm_recovery_services_protection_policy_vm.vm_backup_policy.id
 }
 
+output "file_share_backup_policy_name" {
+  description = "File share Backup policy name"
+  value       = azurerm_backup_policy_file_share.file_share_backup_policy.name
+}
+
+output "file_share_backup_policy_id" {
+  description = "File share Backup policy ID"
+  value       = azurerm_backup_policy_file_share.file_share_backup_policy.id
+}