+++ outputs.tf
@@ -21,3 +21,12 @@ output "vm_backup_policy_id" {
   value       = module.azure-backup.vm_backup_policy_id
 }
 
+output "file_share_backup_policy_name" {
+  description = "File share Backup policy name"
+  value       = module.azure-backup.file_share_backup_policy_name
+}
+
+output "file_share_backup_policy_id" {
+  description = "File share Backup policy ID"
+  value       = module.azure-backup.file_share_backup_policy_id
+}